+++ pkg/apis/kai/v1/config_types.go
@@ -10,7 +10,9 @@ import (
 	"github.com/NVIDIA/KAI-scheduler/pkg/apis/kai/v1/node_scale_adjuster"
 	"github.com/NVIDIA/KAI-scheduler/pkg/apis/kai/v1/pod_group_controller"
 	"github.com/NVIDIA/KAI-scheduler/pkg/apis/kai/v1/pod_grouper"
+	"github.com/NVIDIA/KAI-scheduler/pkg/apis/kai/v1/prometheus"
 	"github.com/NVIDIA/KAI-scheduler/pkg/apis/kai/v1/queue_controller"
+	"github.com/NVIDIA/KAI-scheduler/pkg/apis/kai/v1/scheduler"
 	"github.com/NVIDIA/KAI-scheduler/pkg/common/constants"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -59,6 +61,10 @@ type ConfigSpec struct {
 	// +kubebuilder:validation:Optional
 	Admission *admission.Admission `json:"admission,omitempty"`
 
+	// Scheduler is the configuration for the scheduler microservice
+	// +kubebuilder:validation:Optional
+	Scheduler *scheduler.Scheduler `json:"scheduler,omitempty"`
+
 	// QueueController specifies configuration for the queue controller
 	// +kubebuilder:validation:Optional
 	QueueController *queue_controller.QueueController `json:"queueController,omitempty"`
@@ -70,6 +76,10 @@ type ConfigSpec struct {
 	// NodeScaleAdjuster specifies configuration for the node-scale-adjuster
 	// +kubebuilder:validation:Optional
 	NodeScaleAdjuster *node_scale_adjuster.NodeScaleAdjuster `json:"nodeScaleAdjuster,omitempty"`
+
+	// Prometheus specifies configuration for Prometheus monitoring
+	// +kubebuilder:validation:Optional
+	Prometheus *prometheus.Prometheus `json:"prometheus,omitempty"`
 }
 
 func (c *ConfigSpec) SetDefaultsWhereNeeded() {
@@ -88,6 +98,9 @@ func (c *ConfigSpec) SetDefaultsWhereNeeded() {
 	c.PodGrouper = common.SetDefault(c.PodGrouper, &pod_grouper.PodGrouper{})
 	c.PodGrouper.SetDefaultsWhereNeeded(c.Global.ReplicaCount)
 
+	c.Scheduler = common.SetDefault(c.Scheduler, &scheduler.Scheduler{})
+	c.Scheduler.SetDefaultsWhereNeeded(c.Global.ReplicaCount)
+
 	c.PodGroupController = common.SetDefault(c.PodGroupController, &pod_group_controller.PodGroupController{})
 	c.PodGroupController.SetDefaultsWhereNeeded(c.Global.ReplicaCount)
 
@@ -96,6 +109,9 @@ func (c *ConfigSpec) SetDefaultsWhereNeeded() {
 
 	c.NodeScaleAdjuster = common.SetDefault(c.NodeScaleAdjuster, &node_scale_adjuster.NodeScaleAdjuster{})
 	c.NodeScaleAdjuster.SetDefaultsWhereNeeded()
+
+	c.Prometheus = common.SetDefault(c.Prometheus, &prometheus.Prometheus{})
+	c.Prometheus.SetDefaultsWhereNeeded()
 }
 
 // ConfigStatus defines the observed state of Config